Moved redirect state to folders reducer. Store it as a URL instead of boolean.
