+++ src/js/messaging/reducers/folders.js
@@ -1,12 +1,20 @@
 import set from 'lodash/fp/set';
 import concat from 'lodash/fp/concat';
 
+import { paths } from '../config';
+
 import {
   CREATE_FOLDER_SUCCESS,
+  DELETE_COMPOSE_MESSAGE,
   DELETE_FOLDER_SUCCESS,
+  DELETE_MESSAGE_SUCCESS,
   FETCH_FOLDER_SUCCESS,
   FETCH_FOLDERS_SUCCESS,
   LOADING_FOLDER,
+  MOVE_MESSAGE_SUCCESS,
+  RESET_REDIRECT,
+  SAVE_DRAFT_SUCCESS,
+  SEND_MESSAGE_SUCCESS,
   SET_CURRENT_FOLDER,
   TOGGLE_FOLDER_NAV,
   TOGGLE_MANAGED_FOLDERS
@@ -37,7 +45,8 @@ const initialState = {
     nav: {
       foldersExpanded: false,
       visible: false
-    }
+    },
+    redirect: null
   }
 };
 
@@ -98,6 +107,21 @@ export default function folders(state = initialState, action) {
     case SET_CURRENT_FOLDER:
       // The + forces +action.folderId to be a number
       return set('data.currentItem.persistFolder', +action.folderId, state);
+
+    case DELETE_COMPOSE_MESSAGE:
+    case DELETE_MESSAGE_SUCCESS:
+    case MOVE_MESSAGE_SUCCESS:
+    case SAVE_DRAFT_SUCCESS:
+    case SEND_MESSAGE_SUCCESS: {
+      // Upon completing any of these actions,
+      // set the redirect to the most recent folder.
+      const lastFolderId = state.data.currentItem.persistFolder;
+      const url = `${paths.FOLDERS_URL}/${lastFolderId}`;
+      return set('ui.redirect', url, state);
+    }
+
+    case RESET_REDIRECT:
+      return set('ui.redirect', null, state);
 
     default:
       return state;
